Etc: PR 리뷰 반영 - 프로토콜 ClassNameRenderable 파일 분리
