+++ Pyonsnal-Color/Pyonsnal-Color/Extension/NSObject+.swift
@@ -7,14 +7,4 @@
 
 import Foundation
 
-protocol ClassNameRenderable {
-    static var className: String { get }
-}
-
-extension ClassNameRenderable {
-    static var className: String {
-        return String(describing: self)
-    }
-}
-
 extension NSObject: ClassNameRenderable {}
